LayoutTests/http/wpt/webauthn/public-key-credential-create-success-local.https-expected.txt

CONSOLE MESSAGE: User gesture is not detected. To use the WebAuthn API, call 'navigator.credentials.create' or 'navigator.credentials.get' within user activated events.
CONSOLE MESSAGE: User gesture is not detected. To use the WebAuthn API, call 'navigator.credentials.create' or 'navigator.credentials.get' within user activated events.
CONSOLE MESSAGE: User gesture is not detected. To use the WebAuthn API, call 'navigator.credentials.create' or 'navigator.credentials.get' within user activated events.
CONSOLE MESSAGE: User gesture is not detected. To use the WebAuthn API, call 'navigator.credentials.create' or 'navigator.credentials.get' within user activated events.
CONSOLE MESSAGE: User gesture is not detected. To use the WebAuthn API, call 'navigator.credentials.create' or 'navigator.credentials.get' within user activated events.
CONSOLE MESSAGE: User gesture is not detected. To use the WebAuthn API, call 'navigator.credentials.create' or 'navigator.credentials.get' within user activated events.
CONSOLE MESSAGE: User gesture is not detected. To use the WebAuthn API, call 'navigator.credentials.create' or 'navigator.credentials.get' within user activated events.
CONSOLE MESSAGE: User gesture is not detected. To use the WebAuthn API, call 'navigator.credentials.create' or 'navigator.credentials.get' within user activated events.

PASS PublicKeyCredential's [[create]] with minimum options in a mock local authenticator.
PASS PublicKeyCredential's [[create]] with authenticatorSelection { 'platform' } in a mock local authenticator.
PASS PublicKeyCredential's [[create]] with matched exclude credential ids but not transports in a mock local authenticator.
PASS PublicKeyCredential's [[create]] with none attestation in a mock local authenticator.
PASS PublicKeyCredential's [[create]] with indirect attestation in a mock local authenticator.
PASS PublicKeyCredential's [[create]] with direct attestation in a mock local authenticator.
PASS PublicKeyCredential's [[create]] with duplicate credential in a mock local authenticator.
PASS PublicKeyCredential's [[create]] with duplicate credential in a mock local authenticator. 2
PASS PublicKeyCredential's [[create]] with user presence in a mock local authenticator.
